Move test to folders
- Разбил тесты из SmokeTest в отдельные папка (по страницам)
- Добавил дополнительные тесты
