+++ SwiftRadioUITests/Pages/CommonPage.swift
@@ -19,6 +19,18 @@ class CommonPage: XCTest {
         app.navigationBars.buttons["Back"]
     }
     
+    var navigationBarTitle: XCUIElement {
+        app.navigationBars.element.staticTexts.element(boundBy: 0)
+    }
+    
+    var navigationBarBackButton: XCUIElement{
+        app.navigationBars.buttons.element(boundBy: 0)
+    }
+    
+    func tapNavigationBarBackButton() {
+        navigationBarBackButton.tap()
+    }
+    
     func tapNavigationBackButton() {
         navigationBar.tap()
     }
