+++ SwiftRadioUITests/Tests/MainPageTest.swift
@@ -10,6 +10,12 @@ import XCTest
 
 final class MainPageTest: BaseTest {
     
+    // Smoke. Список радиостанций отображается при открытии приложения
+    func testListOfStationsIsShownOnAppLaunch() {
+        pages.mainPage()
+            .waitForFirstCellToAppear()
+    }
+    
     // Bottom bar плеер становится активным после запуска радио станции
     func testNowPlayingBottomBarIsEnabled() {
         pages.mainPage()
@@ -26,7 +32,7 @@ final class MainPageTest: BaseTest {
     func testPulltoRefreshStations() {
         pages.mainPage()
             .waitForFirstCellToAppear()
-            .PullToRefresh()
+            .pullToRefresh()
             .waitForFirstCellToAppear()
     }
